# subscription/subscription.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "subscription/subscription.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "SubscriptionService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/subscriptions": {
      "get": {
        "operationId": "SubscriptionService_GetSubscriptionsByUserID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/subscription_serviceGetSubscriptionsByIDResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "UserID",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "SubscriptionService"
        ]
      },
      "post": {
        "operationId": "SubscriptionService_AddSubscription",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/subscription_serviceAddSubscriptionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/subscription_serviceAddSubscriptionRequest"
            }
          }
        ],
        "tags": [
          "SubscriptionService"
        ]
      }
    },
    "/v1/subscriptions/{UserID}/{ModID}": {
      "delete": {
        "operationId": "SubscriptionService_RemoveSubscription",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/subscription_serviceRemoveSubscriptionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "UserID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "ModID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "SubscriptionService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "subscription_serviceAddSubscriptionRequest": {
      "type": "object",
      "properties": {
        "UserID": {
          "type": "string"
        },
        "ModID": {
          "type": "string"
        }
      },
      "title": "AddSubscription"
    },
    "subscription_serviceAddSubscriptionResponse": {
      "type": "object"
    },
    "subscription_serviceGetSubscriptionsByIDResponse": {
      "type": "object",
      "properties": {
        "Subscriptions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/subscription_serviceSubscription"
          }
        }
      }
    },
    "subscription_serviceRemoveSubscriptionResponse": {
      "type": "object"
    },
    "subscription_serviceSubscription": {
      "type": "object",
      "properties": {
        "UserID": {
          "type": "string"
        },
        "ModID": {
          "type": "string"
        }
      }
    }
  }
}